feat: building out navigation component
