+++ crown-clothing/src/index.scss
@@ -1,0 +1,21 @@
+body {
+  margin: 0;
+  font-family: "Open Sans Condensed", sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
+    monospace;
+}
+
+// Anchor Tags
+a {
+  text-decoration: none;
+  color: black;
+}
+
+* {
+  box-sixing: border-box;
+}
