+++ crown-clothing/src/routes/navigation/navigation.styles.scss
@@ -1,0 +1,23 @@
+.navigation {
+  height: 70px;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 25px;
+  .logo-container {
+    height: 100%;
+    width: 70px;
+    padding: 25px;
+  }
+  .nav-links-container {
+    width: 50%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    .nav-link {
+      padding: 10px 15px;
+      cursor: pointer;
+    }
+  }
+}
